+++ src/QueryBuilder/AbstractDMLQueryBuilder.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\QueryBuilder;
 
-use Generator;
 use JsonException;
 use Yiisoft\Db\Constraint\Constraint;
 use Yiisoft\Db\Constraint\IndexConstraint;
@@ -26,10 +25,8 @@ use function array_map;
 use function array_merge;
 use function array_unique;
 use function array_values;
-use function count;
 use function implode;
 use function in_array;
-use function is_array;
 use function is_string;
 use function json_encode;
 use function preg_match;
@@ -52,28 +49,23 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
     ) {
     }
 
-    public function batchInsert(string $table, array $columns, iterable|Generator $rows, array &$params = []): string
+    public function batchInsert(string $table, array $columns, iterable $rows, array &$params = []): string
     {
         if (empty($rows)) {
             return '';
         }
 
-        if (($tableSchema = $this->schema->getTableSchema($table)) !== null) {
-            $columnSchemas = $tableSchema->getColumns();
-        } else {
-            $columnSchemas = [];
-        }
-
-        $mappedNames = $this->getNormalizeColumnNames($table, $columns);
         $values = [];
+        $columns = $this->getNormalizeColumnNames('', $columns);
+        $columnSchemas = $this->schema->getTableSchema($table)?->getColumns() ?? [];
 
-        /** @psalm-var array<array-key, array<array-key, string>> $rows */
         foreach ($rows as $row) {
+            $i = 0;
             $placeholders = [];
-            foreach ($row as $index => $value) {
-                if (isset($columns[$index], $mappedNames[$columns[$index]], $columnSchemas[$mappedNames[$columns[$index]]])) {
-                    /** @psalm-var mixed $value */
-                    $value = $this->getTypecastValue($value, $columnSchemas[$mappedNames[$columns[$index]]]);
+
+            foreach ($row as $value) {
+                if (isset($columns[$i], $columnSchemas[$columns[$i]])) {
+                    $value = $columnSchemas[$columns[$i]]->dbTypecast($value);
                 }
 
                 if ($value instanceof ExpressionInterface) {
@@ -81,6 +73,8 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
                 } else {
                     $placeholders[] = $this->queryBuilder->bindParam($value, $params);
                 }
+
+                ++$i;
             }
             $values[] = '(' . implode(', ', $placeholders) . ')';
         }
@@ -89,12 +83,12 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
             return '';
         }
 
-        foreach ($columns as $i => $name) {
-            $columns[$i] = $this->quoter->quoteColumnName($mappedNames[$name]);
-        }
+        $columns = array_map(
+            [$this->quoter, 'quoteColumnName'],
+            $columns,
+        );
 
-        return 'INSERT INTO '
-            . $this->quoter->quoteTableName($table)
+        return 'INSERT INTO ' . $this->quoter->quoteTableName($table)
             . ' (' . implode(', ', $columns) . ') VALUES ' . implode(', ', $values);
     }
 
@@ -108,17 +102,11 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
 
     public function insert(string $table, QueryInterface|array $columns, array &$params = []): string
     {
-        /**
-         * @psalm-var string[] $names
-         * @psalm-var string[] $placeholders
-         * @psalm-var string $values
-         */
         [$names, $placeholders, $values, $params] = $this->prepareInsertValues($table, $columns, $params);
 
-        return 'INSERT INTO '
-            . $this->quoter->quoteTableName($table)
+        return 'INSERT INTO ' . $this->quoter->quoteTableName($table)
             . (!empty($names) ? ' (' . implode(', ', $names) . ')' : '')
-            . (!empty($placeholders) ? ' VALUES (' . implode(', ', $placeholders) . ')' : $values);
+            . (!empty($placeholders) ? ' VALUES (' . implode(', ', $placeholders) . ')' : ' ' . $values);
     }
 
     public function insertWithReturningPks(string $table, QueryInterface|array $columns, array &$params = []): string
@@ -133,10 +121,9 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
 
     public function update(string $table, array $columns, array|string $condition, array &$params = []): string
     {
-        /** @psalm-var string[] $lines */
         [$lines, $params] = $this->prepareUpdateSets($table, $columns, $params);
+
         $sql = 'UPDATE ' . $this->quoter->quoteTableName($table) . ' SET ' . implode(', ', $lines);
-        /** @psalm-var array $params */
         $where = $this->queryBuilder->buildWhere($condition, $params);
 
         return $where === '' ? $sql : $sql . ' ' . $where;
@@ -163,20 +150,21 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
      * @throws InvalidConfigException
      * @throws NotSupportedException
      *
-     * @return array Array of column names, values, and params.
+     * @return array Array of quoted column names, values, and params.
+     * @psalm-return array{0: string[], 1: string, 2: array}
      */
     protected function prepareInsertSelectSubQuery(QueryInterface $columns, array $params = []): array
     {
-        if (empty($columns->getSelect()) || in_array('*', $columns->getSelect(), true)) {
+        /** @psalm-var string[] $select */
+        $select = $columns->getSelect();
+
+        if (empty($select) || in_array('*', $select, true)) {
             throw new InvalidArgumentException('Expected select query object with enumerated (named) parameters');
         }
 
         [$values, $params] = $this->queryBuilder->build($columns, $params);
 
         $names = [];
-        $values = ' ' . $values;
-        /** @psalm-var string[] $select */
-        $select = $columns->getSelect();
 
         foreach ($select as $title => $field) {
             if (is_string($title)) {
@@ -204,37 +192,41 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
      * @throws InvalidConfigException
      * @throws InvalidArgumentException
      * @throws NotSupportedException
+     *
+     * @return array Array of quoted column names, placeholders, values, and params.
+     * @psalm-return array{0: string[], 1: string[], 2: string, 3: array}
      */
     protected function prepareInsertValues(string $table, array|QueryInterface $columns, array $params = []): array
     {
-        $tableSchema = $this->schema->getTableSchema($table);
-        $columnSchemas = $tableSchema !== null ? $tableSchema->getColumns() : [];
-        $names = [];
-        $placeholders = [];
-        $values = ' DEFAULT VALUES';
+        if (empty($columns)) {
+            return [[], [], 'DEFAULT VALUES', []];
+        }
 
         if ($columns instanceof QueryInterface) {
             [$names, $values, $params] = $this->prepareInsertSelectSubQuery($columns, $params);
-        } else {
-            $columns = $this->normalizeColumnNames($table, $columns);
-            /**
-             * @psalm-var mixed $value
-             * @psalm-var array<string, mixed> $columns
-             */
-            foreach ($columns as $name => $value) {
-                $names[] = $this->quoter->quoteColumnName($name);
-                /** @var mixed $value */
-                $value = $this->getTypecastValue($value, $columnSchemas[$name] ?? null);
+            return [$names, [], $values, $params];
+        }
 
-                if ($value instanceof ExpressionInterface) {
-                    $placeholders[] = $this->queryBuilder->buildExpression($value, $params);
-                } else {
-                    $placeholders[] = $this->queryBuilder->bindParam($value, $params);
-                }
+        $names = [];
+        $placeholders = [];
+        $columns = $this->normalizeColumnNames('', $columns);
+        $columnSchemas = $this->schema->getTableSchema($table)?->getColumns() ?? [];
+
+        foreach ($columns as $name => $value) {
+            $names[] = $this->quoter->quoteColumnName($name);
+
+            if (isset($columnSchemas[$name])) {
+                $value = $columnSchemas[$name]->dbTypecast($value);
+            }
+
+            if ($value instanceof ExpressionInterface) {
+                $placeholders[] = $this->queryBuilder->buildExpression($value, $params);
+            } else {
+                $placeholders[] = $this->queryBuilder->bindParam($value, $params);
             }
         }
 
-        return [$names, $placeholders, $values, $params];
+        return [$names, $placeholders, '', $params];
     }
 
     /**
@@ -244,21 +236,20 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
      * @throws InvalidConfigException
      * @throws InvalidArgumentException
      * @throws NotSupportedException
+     *
+     * @psalm-return array{0: string[], 1: array}
      */
     protected function prepareUpdateSets(string $table, array $columns, array $params = []): array
     {
-        $tableSchema = $this->schema->getTableSchema($table);
-        $columnSchemas = $tableSchema !== null ? $tableSchema->getColumns() : [];
         $sets = [];
-        $columns = $this->normalizeColumnNames($table, $columns);
+        $columns = $this->normalizeColumnNames('', $columns);
+        $columnSchemas = $this->schema->getTableSchema($table)?->getColumns() ?? [];
 
-        /**
-         * @psalm-var array<string, mixed> $columns
-         * @psalm-var mixed $value
-         */
         foreach ($columns as $name => $value) {
-            /** @psalm-var mixed $value */
-            $value = isset($columnSchemas[$name]) ? $columnSchemas[$name]->dbTypecast($value) : $value;
+            if (isset($columnSchemas[$name])) {
+                $value = $columnSchemas[$name]->dbTypecast($value);
+            }
+
             if ($value instanceof ExpressionInterface) {
                 $placeholder = $this->queryBuilder->buildExpression($value, $params);
             } else {
@@ -272,7 +263,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
     }
 
     /**
-     * Prepare column names and placeholders for "upsert" operation.
+     * Prepare column names and constraints for "upsert" operation.
      *
      * @throws Exception
      * @throws InvalidArgumentException
@@ -280,7 +271,11 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
      * @throws JsonException
      * @throws NotSupportedException
      *
+     * @psalm-param array<string, mixed>|QueryInterface $insertColumns
      * @psalm-param Constraint[] $constraints
+     *
+     * @return array Array of unique, insert and update quoted column names.
+     * @psalm-return array{0: string[], 1: string[], 2: string[]|null}
      */
     protected function prepareUpsertColumns(
         string $table,
@@ -288,42 +283,28 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
         QueryInterface|bool|array $updateColumns,
         array &$constraints = []
     ): array {
-        $insertNames = [];
-
-        if (!$insertColumns instanceof QueryInterface) {
-            $insertColumns = $this->normalizeColumnNames($table, $insertColumns);
-        }
-
-        if (is_array($updateColumns)) {
-            $updateColumns = $this->normalizeColumnNames($table, $updateColumns);
-        }
-
         if ($insertColumns instanceof QueryInterface) {
-            /** @psalm-var list<string> $insertNames */
             [$insertNames] = $this->prepareInsertSelectSubQuery($insertColumns);
         } else {
-            /** @psalm-var array<string, string> $insertColumns */
-            foreach ($insertColumns as $key => $_value) {
-                $insertNames[] = $this->quoter->quoteColumnName($key);
-            }
+            $insertNames = $this->getNormalizeColumnNames('', array_keys($insertColumns));
+
+            $insertNames = array_map(
+                [$this->quoter, 'quoteColumnName'],
+                $insertNames,
+            );
         }
 
-        /** @psalm-var string[] $uniqueNames */
         $uniqueNames = $this->getTableUniqueColumnNames($table, $insertNames, $constraints);
 
-        foreach ($uniqueNames as $key => $name) {
-            $insertNames[$key] = $this->quoter->quoteColumnName($name);
+        if ($updateColumns === true) {
+            return [$uniqueNames, $insertNames, array_diff($insertNames, $uniqueNames)];
         }
 
-        if ($updateColumns !== true) {
-            return [$uniqueNames, $insertNames, null];
-        }
-
-        return [$uniqueNames, $insertNames, array_diff($insertNames, $uniqueNames)];
+        return [$uniqueNames, $insertNames, null];
     }
 
     /**
-     * Returns all column names belonging to constraints enforcing uniqueness (`PRIMARY KEY`, `UNIQUE INDEX`, etc.)
+     * Returns all quoted column names belonging to constraints enforcing uniqueness (`PRIMARY KEY`, `UNIQUE INDEX`, etc.)
      * for the named table removing constraints which didn't cover the specified column list.
      *
      * The column list will be unique by column names.
@@ -335,7 +316,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
      *
      * @throws JsonException
      *
-     * @return array The column list.
+     * @return string[] The quoted column names.
      *
      * @psalm-param Constraint[] $constraints
     */
@@ -365,9 +346,8 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
          */
         $constraints = array_combine(
             array_map(
-                static function (Constraint $constraint) {
-                    $columns = $constraint->getColumnNames() ?? [];
-                    $columns = is_array($columns) ? $columns : [$columns];
+                static function (Constraint $constraint): string {
+                    $columns = (array) $constraint->getColumnNames();
                     sort($columns, SORT_STRING);
                     return json_encode($columns, JSON_THROW_ON_ERROR);
                 },
@@ -383,18 +363,16 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
         $constraints = array_values(
             array_filter(
                 $constraints,
-                static function (Constraint $constraint) use ($quoter, $columns, &$columnNames) {
-                    /** @psalm-var string[]|string $getColumnNames */
-                    $getColumnNames = $constraint->getColumnNames() ?? [];
-                    $constraintColumnNames = [];
+                static function (Constraint $constraint) use ($quoter, $columns, &$columnNames): bool {
+                    /** @psalm-var string[] $constraintColumnNames */
+                    $constraintColumnNames = (array) $constraint->getColumnNames();
 
-                    if (is_array($getColumnNames)) {
-                        foreach ($getColumnNames as $columnName) {
-                            $constraintColumnNames[] = $quoter->quoteColumnName($columnName);
-                        }
-                    }
+                    $constraintColumnNames = array_map(
+                        [$quoter, 'quoteColumnName'],
+                        $constraintColumnNames,
+                    );
 
-                    $result = !array_diff($constraintColumnNames, $columns);
+                    $result = empty(array_diff($constraintColumnNames, $columns));
 
                     if ($result) {
                         $columnNames = array_merge((array) $columnNames, $constraintColumnNames);
@@ -405,12 +383,14 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
             )
         );
 
-        /** @psalm-var Constraint[] $columnNames */
+        /** @psalm-var string[] $columnNames */
         return array_unique($columnNames);
     }
 
     /**
      * @return mixed The typecast value of the given column.
+     *
+     * @deprecated will be removed in version 2.0.0
      */
     protected function getTypecastValue(mixed $value, ColumnSchemaInterface $columnSchema = null): mixed
     {
@@ -422,61 +402,39 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
     }
 
     /**
-     * Normalizes the column names for the given table.
+     * Normalizes the column names.
      *
-     * @param string $table The table to save the data into.
-     * @param array $columns The column data (name => value) to save into the table or instance of
-     * {@see QueryInterface} to perform `INSERT INTO ... SELECT` SQL statement. Passing of {@see QueryInterface}.
+     * @param string $table Not used. Could be empty string. Will be removed in version 2.0.0.
+     * @param array $columns The column data (name => value).
      *
      * @return array The normalized column names (name => value).
+     *
+     * @psalm-return array<string, mixed>
      */
     protected function normalizeColumnNames(string $table, array $columns): array
     {
-        /** @var string[] $columnList */
-        $columnList = array_keys($columns);
-        $mappedNames = $this->getNormalizeColumnNames($table, $columnList);
+        /** @var string[] $columnNames */
+        $columnNames = array_keys($columns);
+        $normalizedNames = $this->getNormalizeColumnNames('', $columnNames);
 
-        /** @psalm-var array $normalizedColumns */
-        $normalizedColumns = [];
-
-        /**
-         * @psalm-var string $name
-         * @psalm-var mixed $value
-         */
-        foreach ($columns as $name => $value) {
-            $mappedName = $mappedNames[$name] ?? $name;
-            /** @psalm-var mixed */
-            $normalizedColumns[$mappedName] = $value;
-        }
-
-        return $normalizedColumns;
+        return array_combine($normalizedNames, $columns);
     }
 
     /**
-     * Get a map of normalized columns
+     * Get normalized column names
      *
-     * @param string $table The table to save the data into.
-     * @param string[] $columns The column data (name => value) to save into the table or instance of
-     * {@see QueryInterface} to perform `INSERT INTO ... SELECT` SQL statement. Passing of {@see QueryInterface}.
+     * @param string $table Not used. Could be empty string. Will be removed in version 2.0.0.
+     * @param string[] $columns The column names.
      *
-     * @return string[] Map of normalized columns.
+     * @return string[] Normalized column names.
      */
     protected function getNormalizeColumnNames(string $table, array $columns): array
     {
         $normalizedNames = [];
-        $rawTableName = $this->schema->getRawTableName($table);
 
         foreach ($columns as $name) {
-            $parts = $this->quoter->getTableNameParts($name, true);
-
-            if (count($parts) === 2 && $this->schema->getRawTableName($parts[0]) === $rawTableName) {
-                $normalizedName = $parts[count($parts) - 1];
-            } else {
-                $normalizedName = $name;
-            }
-            $normalizedName = $this->quoter->ensureColumnName($normalizedName);
-
-            $normalizedNames[$name] = $normalizedName;
+            $normalizedName = $this->quoter->ensureColumnName($name);
+            $normalizedNames[] = $this->quoter->unquoteSimpleColumnName($normalizedName);
         }
 
         return $normalizedNames;