+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.1.2 under development
 
+- Enh #746: Refactor `AbstractDMLQueryBuilder` (@Tigrov)
+- Bug #746: Fix `AbstractDMLQueryBuilder::upsert()` when unique index is not at the first position of inserted values (@Tigrov)
+- Bug #746: Typecast values in `AbstractDMLQueryBuilder::batchInsert()` if column names with table name and brackets (@Tigrov)
+- Bug #746, #61: Typecast values in `AbstractDMLQueryBuilder::batchInsert()` if values with string keys (@Tigrov)
+- Enh #746: Enhanced documentation of `batchInsert()` and `update()` methods of `DMLQueryBuilderInterface` interface (@Tigrov)
 - Bug #751: Fix collected debug actions (@xepozz)
 - Chg #755: Deprecate `TableSchemaInterface::compositeForeignKey()` (@Tigrov) 
 - Enh #756: Refactor `Quoter` (@Tigrov) 
@@ -10,6 +15,7 @@
 - Bug #756: Fix `Quoter::quoteTableName()` for sub-query with alias (@Tigrov)
 - Bug #761: Quote aliases of CTE in `WITH` queries (@Tigrov)
 - Chg #765: Deprecate `SchemaInterface::TYPE_JSONB` (@Tigrov)
+- Enh #770: Move methods from concrete `Command` class to `AbstractPdoCommand` class (@Tigrov)
 
 ## 1.1.1 August 16, 2023